Move Myclient into internal helpers
